More work on the SongPropertiesWidget removed float versions of speed and setSpeed in Song class

+++ ui/model/SongListModel.hpp
@@ -27,7 +27,7 @@ public:
 public slots:
     void setRowsPerBeat(int rowsPerBeat);
     void setRowsPerMeasure(int rowsPerMeasure);
-    void setSpeed(double speed);
+    void setSpeed(int speed);
 
 
 protected:
